+++ packages/benchmark-js/test/benchmark.ts
@@ -6,9 +6,7 @@ const benchmark = new BenchmarkGroup();
 benchmark.setEmptyTest();
 benchmark.add('Benchmark-1', () => {});
 benchmark.add('Benchmark-2', (arg1, arg2, arg3) => arg3, {
-    argument: { count: 3 },
-    onGetArguments: () => ({ arguments: ['hello', 'world', '!'] }),
-    onGetArgumentsInPrebenchmarkStage: () => [{ arguments: [1, 2, undefined] }, { arguments: [true, () => {}, 'str'] }],
+    args: [[1, 2], [3], [5, 6]],
 });
 benchmark.writeTestersCode();
 benchmark.run();