+++ packages/benchmark-js/CHANGELOG.md
@@ -5,8 +5,9 @@
 - Improve the accuracy of the minimum duration time.
 - Add method `BenchmarkGroup.setEmptyTest` to get the maximum ops as a base line.
 - Add methods `Benchmark.writeTesterCode` and `BenchmarkGroup.writeTestersCode` to print the tester source code.
-- Add Support to pass arguments to `testFn`.
+- Add support to pass arguments to `testFn`.
 - Add adjust-benchmarking stage after pre-benchmarking stage but before formal-benchmarking stage.
+- Add support to set selections for each `testFn`'s argument.
 
 ### Bug Fixes